Updating style.css
Adding in the nav ul selector so I can  removes the bullet points from the list. I also made it so that it uses flexbox for horizontal alignment and the justify-content in order to  center the navigation text. I also made it to that it removes default padding and added in a semi-transparent black background.