Updating style.css
I added in these selectors that change the colors of the headers on each page. I did this because i found it visually appealing and it could make it it easier for users to differentiate the pages visually (red for Home, blue for Favorite Genre, and green for My Music). Just the same colors like the Navigation Hover Effect.